+++ app/(client)/layout.tsx
@@ -1,0 +1,15 @@
+import type { Metadata } from 'next'
+
+import { UserLayout } from '@/components/layouts/user-layout'
+
+export const metadata: Metadata = { title: 'Gelda - Client' }
+
+const SubLayout = ({
+  children
+}: Readonly<{
+  children: React.ReactNode
+}>) => {
+  return <UserLayout>{children}</UserLayout>
+}
+
+export default SubLayout
