+++ app/(root)/auth/register/page.tsx
