+++ components/shared/navigation.tsx
@@ -1,10 +1,8 @@
 'use client'
 
-import { useEffect } from 'react'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 import { icons } from 'lucide-react'
-import { isEmpty } from 'lodash'
 
 import { Icon } from '@/components/ui/icon'
 
@@ -17,8 +15,6 @@ import {
 import { Separator } from '@/components/ui/separator'
 
 import { PATH } from '@/lib/constants'
-
-import { useAuth } from '@/hooks/use-auth'
 
 type Menu = {
   label: string
@@ -51,42 +47,32 @@ const menus = [
 
 const Navigation = () => {
   const pathname = usePathname()
-  const { authenticatedUser, loading, authenticate } = useAuth()
-
-  useEffect(() => {
-    authenticate()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
 
   return (
-    <>
-      {!loading && !isEmpty(authenticatedUser) ? (
-        <div className='relative w-full bg-white dark:bg-slate-950'>
-          <NavigationMenu className='w-full px-2'>
-            <NavigationMenuList className='mx-4 gap-6'>
-              {menus.map((menu, idx) => (
-                <NavigationMenuItem key={menu.path}>
-                  <Link href={menu.path} legacyBehavior passHref>
-                    <NavigationMenuLink
-                      className={`py-4 text-sm flex items-center justify-center gap-2 border-black dark:border-white ${
-                        pathname === menu.path
-                          ? 'text-black dark:text-white border-b-2'
-                          : 'text-black dark:text-inherit'
-                      }`}
-                    >
-                      <Icon name={menu.icon as keyof typeof icons} size={14} />
-                      {menu.label}
-                    </NavigationMenuLink>
-                  </Link>
-                </NavigationMenuItem>
-              ))}
-            </NavigationMenuList>
-          </NavigationMenu>
+    <div className='relative w-full bg-white dark:bg-slate-950'>
+      <NavigationMenu className='w-full px-2'>
+        <NavigationMenuList className='mx-4 gap-6'>
+          {menus.map((menu) => (
+            <NavigationMenuItem key={menu.path}>
+              <Link href={menu.path} legacyBehavior passHref>
+                <NavigationMenuLink
+                  className={`py-4 text-sm flex items-center justify-center gap-2 border-black dark:border-white ${
+                    pathname === menu.path
+                      ? 'text-black dark:text-white border-b-2'
+                      : 'text-black dark:text-inherit'
+                  }`}
+                >
+                  <Icon name={menu.icon as keyof typeof icons} size={14} />
+                  {menu.label}
+                </NavigationMenuLink>
+              </Link>
+            </NavigationMenuItem>
+          ))}
+        </NavigationMenuList>
+      </NavigationMenu>
 
-          <Separator />
-        </div>
-      ) : null}
-    </>
+      <Separator />
+    </div>
   )
 }
